requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-generativeai>=0.3.0
